mkall: enforce a uniform interface for operations
Each operation has settings. You instantiate settings and you
use setters to change the desired values. Then you use a method
of the settings to synchronously perform an operation. You get
back the results of the operations. You use getters to access
the desired values in the results objects.

The above user story was already more or less implemented but
this commit enforce more consistency by making sure that we
always call the settings settings and the results results. We
used to have inconsistent naming with plural and singular, and
with request instead of settings, response instead of results.