Rewrite the release publishing process
Take advantage more extensively of Carthage capabilities to
generate a zipped binary that it knows how to install.

While there, also use CocoaPods in a better way so that
downloading the binary with it is also simpler.

Use a better script to reduce the amount of toil entailed
by making a new release for iOS.