Make sure we actually free memory
This diff fixes a serious `facepalm.jpg` bug where the incorrect
method was overriden for releasing memory.

As such, this library was actually not releasing memory.

The `deinit` method is used by Swift. The `dealloc` method is
used by ObjectiveC. We should use `dealloc`.

It is surprising we never noticed. It may perhaps be because most
memory is managed by MK and only a little bit of it is instead
managed through ObjectiveC objects?

I enabled all sanitisers in the previous commit to increase our
confidence that we'll `facepalm.jpg` less in the future.

While there, do not declare methods such as `init` and `deinit`
COUGH COUGH `dealloc` that are supposedly already declared by
`NSObject`. This is an additional measure to making sure we are
really overriding, because otherwise Xcode complains.

I am also sad to note that is probably the most embarrasing
bug that I ever introduced (AFAIK, of course).