Call `[self init]` in initialisers with arguments
This reverts 51290ad22394a98d3b39311510b3383e84a1ab35. I noticed
empirically that MKOrchestra doesn't work at all when we're calling
`[super init]` rather than `[self init]` in initializers that take
arguments. I do not currently have a precise explanation of why that
is the case. Intuitively, we want to call the initializer without
arguments of the current class, not of the parent class?

This, of course, does not imply that we should not call `[super init]`
when we're overriding the argument-less `init`.