Release v0.5.5
This release fixes a very bad bug where we didn't free the memory
we allocated in ObjectiveC. It brings with it a bunch of stuff that
should make such major screwups less likely in the future.

As I mentioned in the related commit, it's interesting also to see
how we never noticed (or we never encountered data indicating the
occurrence of such event, perhaps).